fix(vinyl): force a potential Vinyl file to be a Vinyl file
For some reason, when a Gulp stream passes a "Vinyl file"
Vinyl.isVinyl(vinylFile) fails. Add a `newVinyl = new Vinyl(vinylFile);`
to make sure a Vinyl file is a Vinyl file.

In the future, find out why this is and how to avoid this. This "force"
might cause issues down the road.